move rmq classes to rmq dir and install kafka.
