+++ lib/killjoy.rb
@@ -12,7 +12,6 @@ require "virtus"
 require "yaml"
 
 require "killjoy/after_fork"
-require "killjoy/amqp_configuration"
 require "killjoy/cassandra/blocking_writes_consumer"
 require "killjoy/cassandra/database_cleaner"
 require "killjoy/cassandra/database_configuration"
@@ -25,10 +24,11 @@ require "killjoy/cassandra/writer"
 require "killjoy/consumer"
 require "killjoy/log_line"
 require "killjoy/log_parser"
-require "killjoy/message"
-require "killjoy/message_bus"
 require "killjoy/mongo/consumer"
 require "killjoy/publisher"
+require "killjoy/rmq/amqp_configuration"
+require "killjoy/rmq/message"
+require "killjoy/rmq/message_bus"
 require "killjoy/thread_pool"
 require "killjoy/version"
