+++ lib/killjoy/kafka/message_bus.rb
@@ -1,0 +1,14 @@
+module Killjoy
+  module Kafka
+    class MessageBus
+      def subscribe(consumer)
+      end
+
+      def stop
+      end
+
+      def publish(message)
+      end
+    end
+  end
+end
