+++ lib/killjoy/rmq/amqp_configuration.rb
