+++ lib/killjoy/rmq/message.rb
