+++ lib/killjoy/rmq/message_bus.rb
